refactor: move enrollment module code into enrollment_littlepay
